% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R
\name{create_tournament}
\alias{create_tournament}
\title{Create tournament}
\usage{
create_tournament(tab)
}
\arguments{
\item{tab}{}
}
\value{

}
\description{
Create tournament
}
\examples{
library(bracketr)
create_tournament(worldcup)

}
